requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.1.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
Faker==33.1.0
python-decouple==3.8
google-auth==2.36.0
stripe==11.4.1
pywebpush==2.0.0
